MAINT: Multiptl live agent chats work now...
Had to add a call to `startTopic` after a `topicFinished` message is received. No idea why, but Web and Android do it as well, and it seems to tell the chat service that the system topic is ready to start new vendor and agent conversations. Also added explicit handler for the wait for agent state in the input bar handler, to avoid the debug message (no change in behavior).