STRY5453857: Scrolling to the bottom for New Conversation
We want the topic selection prompt to be scrolled into view when the user ends a conversation, either manually via the New Conversation menu item, or natrally when the topic ends. Also, when an agent chat is starting we want to scroll to show the 'waiting for agent' message.

scrollToBottom method can be used elsewhere where we want to force the bottom of the table to be visible.