MAINT: Addressed image-download 401 issue
When making requests for images we never need to include the X-Require-Logged-In header, as this is just to prevent side-effects in REST API's, which cannot happen with image requests. Including that header means we have to include all fo the cookies as well (when hitting a '/cs' endpoint) since that is how it checks if the user is logged-in. We do not include the cookies for image requests, so we cannot include that header.