STRY5453857: Start Topic flow is improved...
Especially on slow networks, there is much quicker indication that the topic is starting (typing indicator) because we show it _before_ the topic is started, and then replace it with the title when the topic actually starts. Also removes the text-entry sooner, since it cannot be used until the topic starts and an InputText control is presented.